move crud/auth.py to utils/auth.py
