+++ app/routes/auth.py
@@ -1,16 +1,17 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
-from ..schemas import user
+from ..schemas.user import UserLogin, TokenResponse
 from ..crud import user as user_crud
-from ..crud.auth import create_access_token
+from ..utils.auth import create_access_token
 
 router = APIRouter(
     prefix="/api/auth"
+    tags=["auth"]
 )
 
 
-@router.post("/login", response_model=user.TokenResponse)
-def login(login_data: user.UserLogin):
+@router.post("/login", response_model=TokenResponse)
+def login(login_data: UserLogin):
     # temporary login funciton without db and token
     user = user_crud.authenticate_user(login_data.email, login_data.password)
     if not user:
@@ -19,8 +20,8 @@ def login(login_data: user.UserLogin):
     return {"token": token, "user": user}
 
 """
-@router.post("/login", response_model=user.TokenResponse)
-def login(login_data: user.UserLogin, db: Session = Depends(get_db)):
+@router.post("/login", response_model=TokenResponse)
+def login(login_data: UserLogin, db: Session = Depends(get_db)):
     user = user_crud.authenticate_user(db, login_data.email, login_data.password)
     if not user:
         raise HTTPException(status_code=401, detail="Invalid credentials")
